workshops/o11y-vector-ai/inventory-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-http==1.21.0
